Time: 1 ms (45.15%) | Memory: 41.1 MB (71.82%) - LeetSync

+++ 891-score-after-flipping-matrix/score-after-flipping-matrix.java
@@ -36,7 +36,7 @@ class Solution {
         {
             for(int j =0;j<n;j++)
             {
-                sum+= grid[i][j]*(int)Math.pow(2,n-j-1);
+                sum+= grid[i][j]*Math.pow(2,n-j-1);
                 
 
             }
